add parsing test which emulates parsing during indexing
