+++ testData/parser/TypeInBlock.go
@@ -1,0 +1,14 @@
+package exec
+
+func Start() {
+	if c.Process != nil {
+		return 1
+	}
+	type F func(*Cmd) ()
+}
+
+func Start2() {
+	if c.Process != nil {
+		return 1
+	}
+}
