+++ testData/parser/TypeInBlock.txt
@@ -1,0 +1,78 @@
+GO_FILE
+  PACKAGE_CLAUSE
+    PsiElement(package)('package')
+    PsiElement(identifier)('exec')
+  IMPORT_LIST
+    <empty list>
+  FUNCTION_DECLARATION
+    PsiElement(func)('func')
+    PsiElement(identifier)('Start')
+    SIGNATURE
+      PARAMETERS
+        PsiElement(()('(')
+        PsiElement())(')')
+    BLOCK
+      PsiElement({)('{')
+      IF_STATEMENT
+        PsiElement(if)('if')
+        SIMPLE_STATEMENT
+          LEFT_HAND_EXPR_LIST
+            CONDITIONAL_EXPR
+              REFERENCE_EXPRESSION
+                REFERENCE_EXPRESSION
+                  PsiElement(identifier)('c')
+                PsiElement(.)('.')
+                PsiElement(identifier)('Process')
+              PsiElement(!=)('!=')
+              REFERENCE_EXPRESSION
+                PsiElement(identifier)('nil')
+        BLOCK
+          PsiElement({)('{')
+          RETURN_STATEMENT
+            PsiElement(return)('return')
+            LITERAL
+              PsiElement(int)('1')
+          PsiElement(})('}')
+      STATEMENT
+        TYPE_DECLARATION
+          PsiElement(type)('type')
+          TYPE_SPEC
+            SPEC_TYPE
+              PsiElement(identifier)('F')
+              FUNCTION_TYPE
+                PsiElement(func)('func')
+                SIGNATURE
+                  PARAMETERS
+                    PsiElement(()('(')
+                    PARAMETER_DECLARATION
+                      POINTER_TYPE
+                        PsiElement(*)('*')
+                        TYPE
+                          TYPE_REFERENCE_EXPRESSION
+                            PsiElement(identifier)('Cmd')
+                    PsiElement())(')')
+                  RESULT
+                    PARAMETERS
+                      PsiElement(()('(')
+                      PsiElement())(')')
+      PsiElement(})('}')
+  FUNCTION_DECLARATION
+    PsiElement(func)('func')
+    PsiElement(identifier)('Start2')
+    SIGNATURE
+      PARAMETERS
+        PsiElement(()('(')
+        PsiElement())(')')
+    BLOCK
+      PsiElement({)('{')
+      PsiElement(if)('if')
+      PsiElement(identifier)('c')
+      PsiElement(.)('.')
+      PsiElement(identifier)('Process')
+      PsiElement(!=)('!=')
+      PsiElement(identifier)('nil')
+      PsiElement({)('{')
+      PsiElement(return)('return')
+      PsiElement(int)('1')
+      PsiElement(})('}')
+      PsiElement(})('}')
