+++ tests/com/goide/parser/GoParserTestBase.java
@@ -1,0 +1,57 @@
+/*
+ * Copyright 2013-2016 Sergey Ignatov, Alexander Zolotov, Florin Patan
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.goide.parser;
+
+import com.intellij.core.CoreApplicationEnvironment;
+import com.intellij.lang.LanguageExtensionPoint;
+import com.intellij.lang.ParserDefinition;
+import com.intellij.openapi.extensions.Extensions;
+import com.intellij.testFramework.ParsingTestCase;
+import org.jetbrains.annotations.NotNull;
+
+public abstract class GoParserTestBase extends ParsingTestCase {
+  public GoParserTestBase(String dataPath, String fileExt, ParserDefinition... definitions) {super(dataPath, fileExt, definitions);}
+
+  @NotNull
+  @Override
+  protected String getTestDataPath() {
+    return "testData";
+  }
+
+  @Override
+  protected boolean skipSpaces() {
+    return true;
+  }
+
+  @Override
+  protected void doTest(boolean checkErrors) {
+    super.doTest(false);
+    if (checkErrors) {
+      assertFalse(
+        "PsiFile contains error elements",
+        toParseTreeText(myFile, skipSpaces(), includeRanges()).contains("PsiErrorElement")
+      );
+    }
+  }
+
+  @Override
+  protected void setUp() throws Exception {
+    super.setUp();
+    CoreApplicationEnvironment.registerExtensionPoint(
+      Extensions.getRootArea(), "com.intellij.lang.braceMatcher", LanguageExtensionPoint.class);
+  }
+}
